refactor(FoodList) made makeNutritionObj funtion in utils by taking code that adjusted the api results into the object we wanted and turned it into a function
